feat: unify breadcrumbs navigation
