+++ src/components/layouts/AppNavbar/components/index.ts
@@ -1,3 +1,2 @@
 export { default as AnimatedAvatar } from "./AnimatedAvatar.vue";
-export { default as PageBreadcrumb } from "./PageBreadcrumb.vue";
-export { default as WebsiteTitle } from "./WebsiteTitle.vue";
+export { default as BreadCrumbs } from "./BreadCrumbs.vue";
